+++ src/functions.php
@@ -4,4 +4,8 @@ require_once 'PHPHelperId/ribuan.php';
 require_once 'PHPHelperId/re_ribuan.php';
 require_once 'PHPHelperId/romawi.php';
 require_once 'PHPHelperId/rupiah.php';
+require_once 'PHPHelperId/re_rupiah.php';
 require_once 'PHPHelperId/terbilang.php';
+require_once 'PHPHelperId/re_terbilang.php';
+require_once 'PHPHelperId/tertanggal.php';
+require_once 'PHPHelperId/re_tertanggal.php';