+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Semua perubahan penting pada proyek ini akan didokumentasikan dalam file ini.
 Formatnya didasarkan pada [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 dan proyek ini menganut [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2020-03-04
+### Added
+- function `terbilang` mendukung angka minus dan koma
+- test `terbilang` bagian angka minus dan koma
+- function `re_rupiah`, `re_terbilang`, `tertanggal`, `re_tertanggal`
+- test `re_rupiah`, `re_terbilang`, `tertanggal`, `re_tertanggal`
+
 ## [0.0.1] - 2020-02-25
 ### Added
 - function `ribuan`, `re_ribuan`, `romawi`, `rupiah`, `terbilang`